Convert all event listener files
