--- src/events/error/index.js
@@ -1,7 +1,0 @@
-const { sendErrorLog } = require('../../utils/helpers');
-
-module.exports = ({ yagi }) => {
-  yagi.on('error', (error) => {
-    sendErrorLog(yagi, error);
-  });
-};
