+++ src/events/error/index.ts
@@ -1,0 +1,8 @@
+import { sendErrorLog } from '../../utils/helpers';
+import { EventModule } from '../events';
+
+export default function ({ yagi }: EventModule) {
+  yagi.on('error', (error: Error) => {
+    sendErrorLog(yagi, error);
+  });
+}
